Fix formatting of info message
