+++ plugin/clang/utils.py
@@ -251,32 +251,3 @@ class ClangUtils:
             result += method_cursor.brief_comment + "</b>"
 
         return result
-
-    @staticmethod
-    def cleanup_comment(raw_comment):
-        """Cleanup raw doxygen comment."""
-        def pop_prepending_empty_lines(lines):
-            first_non_empty_line_idx = 0
-            for line in lines:
-                if line == '':
-                    first_non_empty_line_idx += 1
-                else:
-                    break
-            return lines[first_non_empty_line_idx:]
-
-        import string
-        lines = raw_comment.split('\n')
-        chars_to_strip = '/' + '*' + string.whitespace
-        lines = [line.lstrip(chars_to_strip) for line in lines]
-        lines = pop_prepending_empty_lines(lines)
-        clean_lines = []
-        is_brief_comment = True
-        for line in lines:
-            if line == '' and is_brief_comment:
-                # Skip lines that belong to brief comment.
-                is_brief_comment = False
-                continue
-            if is_brief_comment:
-                continue
-            clean_lines.append(line)
-        return '<br>'.join(clean_lines)
